QT support added for reading data from image

+++ ImageHandler.h
@@ -2,7 +2,7 @@
 
 struct rgbData
 {
-	uint8_t r, g, b;
+    uint8_t r, g, b;
 };
 
 class Bitmap
@@ -18,7 +18,7 @@ public:
 	void AssignPixelData(uint8_t* pixelData);
 	void DrawRandomImage();
 	void SaveImage();
-	static unsigned char* ReadImage(const char* filename, uint32_t& dataSize);
+    static unsigned char* ReadImage(const char* filename, size_t& dataSize);
 
 private:
 	int m_Width, m_Height;
